requirements.txt
djangorestframework==3.15.1
gunicorn==22.0.0
uvicorn==0.29.0